Finished action plan 1

+++ wicket-core/src/main/java/org/apache/wicket/markup/html/debug/PageView.java
@@ -161,7 +161,7 @@ public final class PageView extends Panel
 			{
 				final ComponentData componentData = listItem.getModelObject();
 
-				listItem.add(new Label("row", Long.toString(listItem.getIndex() + 1)));
+				listItem.add(new Label("row", Long.toString((long)listItem.getIndex() + 1)));
 				listItem.add(new Label("path", componentData.path));
 				listItem.add(new Label("size", Bytes.bytes(componentData.size).toString()));
 				listItem.add(new Label("type", componentData.type));
